fix handling of StringList properties
QStringList properties are used in the property browser to emulate
enums. In this implementation, the whole list is the (ordered) list of
possible names, while the first element is the currently selected element.

The current implementation had bugs, notably that it was setting the
current value before refreshing the list of names. Since the first element
in the list encodes the currently selected symbol, it led to having
the code "stutter" a few times. Signals would be triggered when setValue
was called, then again when the enumNames were changed.

Moreover, it made it pretty hard to know what to do with the list: should
we add the currently selected element in the list as well, or just as
first element ? It was also confusing from an UI point of view as the
selected element would always be the first (as opposed to the element
that was selected by the user the last time).