fix transform graph incremental update when both nodes in a new link have parents
This was the last use case not covered by the last patches on
the subject, namely the case where we have something like:

  f1 -> f2
  f3 -> f4

and the user adds f2 -> f4. Or, worse,

  f1 -> f2
  f1 -> f3

and the user adds f3 -> f2. This last case is invalid, but can easily
happen in bigger graphs if the user is editing its transformer
configuration file (as opposed to only load a known transform file).

This commit reshapes the graph to invert the "child" parts, making
sure that it does not create cycles.