Vizkit3DPlugin: fix segfault when setting viz frame
Setting the plugin viz frame from ruby ended in a segfault
because qtruby accidently calls the wrong slot. To avoid
this the slot is renamed.

Furthermore, to avoid a wrong display name after selecting the
the frame in the property browser until the browser lost focus
no signal is emitted.